Backed out changeset 47f3d13ffb7f
Hmm still causing build failures. Shouldn't be we are releasing on monday and I only have another hour to track this down. Back out so others can keep going.

Bug 1009429 - enhancement: Make the algorithm selection in NSS more flexible
This is patches 1 and 2 (patches 3,4,5,&6 are already in, 7 is waiting on these 2).