Bug 1290334 - Only build SSE2 chacha20 implementation with GCC/Clang. r=kaie
The SSE2 code in chacha20_vec.c relies on syntax that only builds with
GCC/Clang and not e.g. MSVC. When building with MSVC and CPU_ARCH is
x86_64, the generic non-SSE2 chacha20.c needs to be used.

Only build chacha20_vec.c when the compiler is GCC or Clang.