Bug 1397990 - Run clang-format 4.0, r=me
These files were being reformatted by clang-format 4.0.  If you make this
change, then both 3.9 (in CI) and 4.0 are happy with the result.  I don't plan
to do this often, but it is a huge help.

Note that the PK11 ECDSA tests are an odd duck, I have more on that coming.

--HG--
branch : NSS_TLS13_DRAFT19_BRANCH