Bug 1398679 - Make cipher specs properly directional, r=ekr
This makes each cipher spec unidirectional.  This is a tiny bit less efficient
in TLS 1.2 and earlier, where some of the material could be shared (a few
pointers essentially) but it is much more efficient for TLS 1.3.

There is now only one variable of each type on the specs.  Up to now, the specs
had two copies of almost everything to support being used for both read and
write.  Now there are separate specs for reading and writing.  We only
duplicate the pointers to the master secret, and the cipher definitions.

This also does away with the backing array that was used to hold two copies of
specs.  Cipher specs are allocated on the heap as they are used and reference
counted, using the same system as is already used for TLS 1.3.

This uses the |direction| attribute that was previously added for TLS 1.3 and
uses that more thoroughly.

--HG--
branch : NSS_TLS13_DRAFT19_BRANCH