Bug 1413786 - TlsHandshakeFilter should handle fragmented messages, r=ekr
This just saves partial records and saves them until the last fragment, where
they are merged together.

This is only necessary because something on trunk caused DSA certificates to
grow enough to cause the ServerKeyExchange message to be fragmented in the
Pre13Datagram/TlsConnectGenericPre13.InvalidDERSignatureFfdhe/* tests.  Better
to be a little more robust than to not work at all.  This should tide us over
until we want to filter handshake messages that exceed the record size.

--HG--
branch : NSS_TLS13_DRAFT19_BRANCH