Bug 1419721 - SECMOD_CanDeleteInternalModule() should return false when build doesn't support FIPS r=franziskus
Summary:
Firefox/PSM uses SECMOD_CanDeleteInternalModule() to check whether FIPS can be
toggled or not. If the NSS build doesn't support FIPS (as is the case with the
one that Mozilla ships) this should always return false.

There is no point in trying to remove the internal module only to fail and then
switch back. We only support removing the internal module to switch between
FIPS and non-FIPS mode.

Reviewers: franziskus

Reviewed By: franziskus

Bug #: 1419721

Differential Revision: https://phabricator.services.mozilla.com/D271