Bug 1423043 - Enable half-close, r?ttaubert
Summary:
TLS 1.3 explicitly changed to allow close_notify on one half of the
connection.  Since SSL, an endpoint was required to send close_notify if it
received close_notify.  The general agreement was that this was a silly
requirement and that we would remove it and allow one side of the connection to
be closed.  This is critical for some protocols that are being moved to use
TLS.

NSS was almost perfect here.  The only problem was that it suppressed the
second close_notify.  I've added a test for that.

Reviewers: ttaubert

Bug #: 1423043

Differential Revision: https://phabricator.services.mozilla.com/D797