Bug 1534468 - Expose ChaCha20 primitive through PKCS#11, r=ekr
Summary:
This adds a "CTR" mode for ChaCha20.  This takes a composite 16 octet "IV",
which is internally decomposed into a nonce and counter.

This operates like a CTR mode cipher on arbitrary input, up to the ChaCha20
limit of 2^32 x 64 octet blocks.  The counter provided is a starting counter and
it is incremented if more than 64 octets of input is provided.

Reviewers: ekr

Tags: #secure-revision

Bug #: 1534468

Differential Revision: https://phabricator.services.mozilla.com/D23060