Bug 1578626 - Remove undefined nullptr decrement, r=keeler
Summary:
This uses uintptr_t to avoid the worst.  It still looks terrible and might trip
static analysis warnings, but the reinterpret_cast should hide that.

This assumes that sizeof(uintptr_t) == sizeof(void*), so I've added an assertion
so that we'll at least fail the test on those systems.  (We could use
GTEST_SKIP instead, but we don't have that in the version of gtest that we use.)

Reviewers: keeler

Tags: #secure-revision

Bug #: 1578626

Differential Revision: https://phabricator.services.mozilla.com/D44937