Bug 1590972 - Use -std=c99 for all C code, r=jcj
This switches to using -std=c99 for compiling all C code.  Previously, we only
enabled this option for lib/freebl and lib/ssl.

For Linux, this means we need to define _DEFAULT_SOURCE to access some of the
functions we use.  On glibc 2.12 (our oldest supported version), we also need to
define _BSD_SOURCE to access these functions.

The only tricky part is dealing with partial C99 implementation in gcc 4.4.
From what I've seen, the only problem is that - in that mode - it doesn't
support nesting of unnamed fields:
https://gcc.gnu.org/onlinedocs/gcc-4.4.7/gcc/Unnamed-Fields.html

This also switches from -std=c++0x to -std=c++11 as the 0x variant, though
identical in meaning, is deprecated.

Differential Revision: https://phabricator.services.mozilla.com/D50421