Bug 1635509 NSS needs to export the System FIPS state. r=kjacob
Internally, NSS uses the system FIPS state to determine if it needs to go into FIPS mode independent of the database FIPS indication. Some applications need to know this value, particularly if the need to know the FIPS state before they call NSS_Init (NSS_IsFIPS() is only valid after init because it depends on the database indicator which is not known until NSS is intialized.

Differential Revision: https://phabricator.services.mozilla.com/D73986