Bug 1642871 - Allow tickets and PHA after resumption, r=kjacobs
The first part of this is fairly simple: we accidentally disabled sending of
session tickets after resumption.

The second part is much less obvious, because the spec is unclear.  This change
takes the interpretation that it is OK to use post-handshake authentication if
the handshake is resumed, but not OK if the handshake is based on a PSK.  (This
is based on a first-principles understanding of resumption being a continuation
of a certificate-based connection rather than a reading of the spec, see the bug
for why the spec appears to be unhelpful on this point.)

This still prohibits the use of post-handshake authentication if an external PSK
was used, but that is more an abundance of caution than anything principled.

Differential Revision: https://phabricator.services.mozilla.com/D77993