Bug 1712184 NSS tools manpages need to be updated to reflect that sqlite is the default database.
This patch does 2 things:

1) update certutil.xml pk12util.xml modutil.xml and signver.xml to reflect the fact
the the sql database is default. Many of these also has examples of specifying
sql:dirname which is now the default. I did not replace them with dbm:dirname since
we don't want to encourage regressing back. The one exception is in the paragraph
explaining how to get to the old database format.

2) I ran make in the diretory to update the .1 and .html files generated from the .xml
files. There are a number of old updates to the .xml files which haven't been picked
up in their corresponding html or man page files. This updates are included in this
patch.

It is really only necessary to review the changes to the .xml files, the rest were
reviewed when their patches were applied.

bob

Differential Revision: https://phabricator.services.mozilla.com/D115658