Bug 1712647 - Split custom extension handling for ECH. r=bbeurdouche,mt
A new function SSL_CallExtensionWriterOnEchInner() allows applications to have
custom extension handlers called separately for CHInner and CHOuter.

This is a little tricky as ECH needs to construct two versions of CHInner: one
compressed and one not.  This just calls the write handler twice in that case.

The other complication is that a handler might make different choices for
CHInner and CHOuter.  This forces us to stop compressing that extension and any
that follow it when that occurs.  In order to ensure that extensions are
consistently placed, we need to track what can be compressed during both
invocations.

I've retained the quirk where the extensions are built twice.  That might be
something that can be removed in future, but for now it creates a negative
externality that I've noted in documentation.

Differential Revision: https://phabricator.services.mozilla.com/D115852